+++ extensions/tategaki/CHANGELOG.md
@@ -1,3 +1,7 @@
 # tategaki Changelog
 
-## [Initial Version] - {PR_MERGE_DATE}
+## [Initial Release] - {PR_MERGE_DATE}
+
+- Added vertical text conversion with grapheme-level Unicode support
+- Three separator modes: none, space, tab
+- Input from selected text or clipboard